refactor Tracer
update alpaka to 1.2.0
rename SimpleTracer to MegaKernelTracer
refactor MegaKernelTracer, cleaner and compact code
move contents of Platform.h, Accelerator.h, Scan.h, Gather.h into
MegaKernelTracer
remove unused file RAY-Core.h
remove fallback to serial accelerator backend
simplify DeviceConfig

+++ Intern/rayx-core/src/Tracer/DeviceTracer.h
@@ -24,7 +24,7 @@ class RAYX_API DeviceTracer {
   public:
     virtual ~DeviceTracer() = default;
 
-    virtual BundleHistory trace(const Group&, Sequential sequential, uint64_t max_batch_size, int THREAD_COUNT = 1, uint32_t maxEvents = 1) = 0;
+    virtual BundleHistory trace(const Group&, Sequential sequential, int max_batch_size, int THREAD_COUNT = 1, int maxEvents = 1) = 0;
 };
 
 }  // namespace RAYX
